add chat template support and remove the VectorSpace trait
