+++ interfaces/kalosm/src/lib.rs
@@ -9,7 +9,6 @@ pub mod language {
     #![doc = include_str!("../docs/language.md")]
     pub use kalosm_common::ModelLoadingProgress;
     pub use kalosm_common::{accelerated_device_if_available, FileSource};
-    pub use kalosm_language::chat::*;
     pub use kalosm_language::context::*;
     pub use kalosm_language::kalosm_language_model::{
         Embedder as _, EmbedderExt as _, Model as _, ModelExt as _, *,
