+++ interfaces/language-model/Cargo.toml
@@ -13,33 +13,33 @@ futures-util = "0.3.28"
 llm-samplers = { workspace = true }
 log = "0.4.17"
 rand = "0.8.5"
-tokio = { version = "1.28.1", features = ["sync"] }
 serde = { version = "1.0.163", features = ["derive"], optional = true }
 once_cell = "1.18.0"
 tracing = "0.1.37"
-async-openai = { version = "0.24.0", optional = true }
-async-trait = "0.1.73"
 candle-core.workspace = true
 kalosm-sample = { workspace = true }
 kalosm-common.workspace = true
 kalosm-streams.workspace = true
-rayon = "1.10.0"
-postcard = { version = "1.0.8", features = ["use-std"], optional = true }
 thiserror.workspace = true
 lru = { version = "0.12.3", optional = true }
 safetensors = { version = "0.4.3", optional = true }
 tokenizers = { workspace = true }
+minijinja = { version = "2.5.0", features = ["json", "loader"] }
+minijinja-contrib = { version = "2.5.0", features = ["pycompat"] }
+reqwest = "0.12.12"
 
 [dev-dependencies]
 tokio = { version = "1.28.1", features = ["full"] }
 kalosm = { workspace = true, features = ["language"] }
 kalosm-learning = { workspace = true }
+pretty_assertions = "1.4.1"
+postcard = { version = "1.0.8", features = ["use-std"] }
 
 [features]
 default = ["cache"]
-remote = ["async-openai"]
+remote = []
 serde = ["dep:serde", "safetensors"]
-cache = ["serde", "dep:postcard", "dep:lru"]
+cache = ["serde", "dep:lru"]
 
 [package.metadata.docs.rs]
 # Features to pass to Cargo (default: [])
