+++ interfaces/language-model/src/model/generation_parameters.rs
@@ -1,0 +1,214 @@
+use llm_samplers::configure::SamplerChainBuilder;
+use llm_samplers::prelude::*;
+
+/// Parameters to use when generating text.
+#[derive(Debug, Clone, PartialEq)]
+pub struct GenerationParameters {
+    pub(crate) temperature: f32,
+    pub(crate) tau: f32,
+    pub(crate) eta: f32,
+    pub(crate) mu: f32,
+    pub(crate) repetition_penalty: f32,
+    pub(crate) repetition_penalty_range: u32,
+    pub(crate) max_length: u32,
+    pub(crate) stop_on: Option<String>,
+}
+
+impl Default for GenerationParameters {
+    fn default() -> Self {
+        Self {
+            temperature: 0.8,
+            eta: 0.1,
+            tau: 5.,
+            mu: 10.,
+            repetition_penalty: 1.3,
+            repetition_penalty_range: 64,
+            max_length: 128,
+            stop_on: None,
+        }
+    }
+}
+
+impl GenerationParameters {
+    /// Create a sampler chain from the generation parameters.
+    pub fn sampler(self) -> SamplerChain {
+        use llm_samplers::configure::SamplerSlot;
+        let GenerationParameters {
+            temperature,
+            tau,
+            eta,
+            mu,
+            repetition_penalty,
+            repetition_penalty_range,
+            max_length: _,
+            stop_on: _,
+        } = self;
+        SamplerChainBuilder::from([
+            (
+                "repetition",
+                SamplerSlot::new_static(move || {
+                    Box::new(
+                        SampleRepetition::default()
+                            .penalty(repetition_penalty)
+                            .last_n(repetition_penalty_range as usize),
+                    )
+                }),
+            ),
+            (
+                "freqpresence",
+                SamplerSlot::new_static(move || Box::new(SampleFreqPresence::default().last_n(64))),
+            ),
+            (
+                "seqrepetition",
+                SamplerSlot::new_static(move || Box::<SampleSeqRepetition>::default()),
+            ),
+            (
+                "temperature",
+                SamplerSlot::new_static(move || {
+                    Box::new(SampleTemperature::default().temperature(temperature))
+                }),
+            ),
+            (
+                "mirostat2",
+                SamplerSlot::new_static(move || {
+                    Box::new(SampleMirostat2::default().tau(tau).eta(eta).mu(mu))
+                }),
+            ),
+        ])
+        .into_chain()
+    }
+
+    /// Get the mirostat2 sampler from the generation parameters.
+    pub fn mirostat2_sampler(self) -> SampleMirostat2 {
+        SampleMirostat2::default()
+            .tau(self.tau)
+            .eta(self.eta)
+            .mu(self.mu)
+    }
+
+    /// Create a sampler chain from the generation parameters without removing any tokens. This can be useful in combination with [`ModelExt::stream_structured_text_with_sampler`] which may pick unlikely tokens.
+    pub fn bias_only_sampler(self) -> SamplerChain {
+        use llm_samplers::configure::SamplerSlot;
+        let GenerationParameters {
+            temperature,
+            repetition_penalty,
+            repetition_penalty_range,
+            ..
+        } = self;
+        SamplerChainBuilder::from([
+            (
+                "repetition",
+                SamplerSlot::new_static(move || {
+                    Box::new(
+                        SampleRepetition::default()
+                            .penalty(repetition_penalty)
+                            .last_n(repetition_penalty_range as usize),
+                    )
+                }),
+            ),
+            (
+                "freqpresence",
+                SamplerSlot::new_static(move || Box::new(SampleFreqPresence::default().last_n(64))),
+            ),
+            (
+                "seqrepetition",
+                SamplerSlot::new_static(move || Box::<SampleSeqRepetition>::default()),
+            ),
+            (
+                "temperature",
+                SamplerSlot::new_static(move || {
+                    Box::new(SampleTemperature::default().temperature(temperature))
+                }),
+            ),
+        ])
+        .into_chain()
+    }
+
+    /// Set the temperature to use when generating text.
+    pub fn with_temperature(mut self, temperature: f32) -> Self {
+        self.temperature = temperature;
+        self
+    }
+
+    /// Set the tau to use when generating text.
+    pub fn with_tau(mut self, tau: f32) -> Self {
+        self.tau = tau;
+        self
+    }
+
+    /// Set the eta to use when generating text.
+    pub fn with_eta(mut self, eta: f32) -> Self {
+        self.eta = eta;
+        self
+    }
+
+    /// Set the mu to use when generating text.
+    pub fn with_mu(mut self, mu: f32) -> Self {
+        self.mu = mu;
+        self
+    }
+
+    /// Set the repetition penalty to use when generating text.
+    pub fn with_repetition_penalty(mut self, repetition_penalty: f32) -> Self {
+        self.repetition_penalty = repetition_penalty;
+        self
+    }
+
+    /// Set the repetition penalty range to use when generating text.
+    pub fn with_repetition_penalty_range(mut self, repetition_penalty_range: u32) -> Self {
+        self.repetition_penalty_range = repetition_penalty_range;
+        self
+    }
+
+    /// Set the maximum length to use when generating text.
+    pub fn with_max_length(mut self, max_length: u32) -> Self {
+        self.max_length = max_length;
+        self
+    }
+
+    /// Set the string to stop on when generating text.
+    pub fn with_stop_on(mut self, stop_on: impl Into<Option<String>>) -> Self {
+        self.stop_on = stop_on.into();
+        self
+    }
+
+    /// Get the temperature to use when generating text.
+    pub fn temperature(&self) -> f32 {
+        self.temperature
+    }
+
+    /// Get the tau to use when generating text.
+    pub fn tau(&self) -> f32 {
+        self.tau
+    }
+
+    /// Get the eta to use when generating text.
+    pub fn eta(&self) -> f32 {
+        self.eta
+    }
+
+    /// Get the mu to use when generating text.
+    pub fn mu(&self) -> f32 {
+        self.mu
+    }
+
+    /// Get the repetition penalty to use when generating text.
+    pub fn repetition_penalty(&self) -> f32 {
+        self.repetition_penalty
+    }
+
+    /// Get the repetition penalty range to use when generating text.
+    pub fn repetition_penalty_range(&self) -> u32 {
+        self.repetition_penalty_range
+    }
+
+    /// Get the maximum length to use when generating text.
+    pub fn max_length(&self) -> u32 {
+        self.max_length
+    }
+
+    /// Get the string to stop on when generating text.
+    pub fn stop_on(&self) -> Option<&str> {
+        self.stop_on.as_deref()
+    }
+}
