+++ models/kalosm-llama/src/token_stream.rs
